Add a win-bash to support bash shell for MS Windows system.
The win-bash[1] is a Windows port of the famous GNU bash shell. Unlike other
bash ports for Windows(e.g. the cygwin), the win-bash needs no special
environment or DLLs. There is just one binary and that's it.

After this patch, we can write a single consistent script without forking to the
platform when we run any script in this project.

[1] https://sourceforge.net/projects/win-bash