Apply sourcemap to server (#301)
This patch provides source map support for stack traces in node.

It uses the source-map module to replace the paths and line numbers of
source-mapped files with their original paths and line numbers. The output
mimics node's stack trace format with the goal of making every compile-to-JS
language more of a first-class citizen. Source maps are completely general
(not specific to any one language) so you can use source maps with multiple
compile-to-JS languages in the same node process.