Implement Application class.
The Application class is a wrapper class to hide server backend implementation.
In our Absolute project, we are currently using the Express but might want to
change the backend to other things. Even if so, we can change them smoothly.
This change is introducing a new way to route requests. We can refer to the
related example in server/example/router/example.router.js.