Absolute v2: Start from scratch.
There was a previous incarnation of this project. We had written it in
the most quick and dirty manner that was handy because we required it
as a dependency in an experiment. But it made maintenance difficult
eventually.

So, we decided to restart from scratch to build on solid foundations
after long discussion. The following things are important differences
between v1 and v2.
  * Use TypeScript instead of Javascript with Babel.
    - It makes our build system simpler and we can use many syntax
      sugars such as access modifiers in language level.
  * SHOULD write tests in any change.
    - We did skip too many tests in feature implementation. Our goal
      is keeping our test coverage over 85% in v2.
  * SHOULD write design documents when implementing a new feature.
    - We did miss too many documents in feature implementation. It made
      maintenance difficult eventually. It is mandatory to write a
      design document when implementing a new feature in v2.