Apply tslint for linting TypeScript.
After this patch, we can use the following commands:
  $ ./absolute lint
  $ ./absolute lint:fix

Also, this patch is fixing some lint errors that we can fix using
'lint:fix' automatically. On the other hand, the unresolved errors will
be fixed step by step in follow-up patch. Once resoloving the remaining
errors, we will be able to integrate the lint test with CI tools.

ISSUE=#465